+++ src/app/(provider)/(root)/recommend/_components/BottomSheet/BottomSheetRating.tsx
@@ -2,15 +2,15 @@ import { getIconPath } from "@/components/commons/Icon/getIconPath";
 import { ICON } from "@/constants/icon";
 import Image from "next/image";
 
-type ReviewBottomSheetRating = {
+type BottomSheetRatingProps = {
   rating: number;
   handleRatingClick: (rating: number, idx: number) => void;
 };
 
-function ReviewBottomSheetRating({
+function BottomSheetRating({
   rating,
   handleRatingClick,
-}: ReviewBottomSheetRating) {
+}: BottomSheetRatingProps) {
   return (
     <div className="w-full h-10 flex justify-center gap-x-3 aspect-square">
       {Array.from({ length: 5 }).map((_, idx) => (
@@ -32,4 +32,4 @@ function ReviewBottomSheetRating({
   );
 }
 
-export default ReviewBottomSheetRating;
+export default BottomSheetRating;
