--- src/app/(provider)/(root)/recommend/_components/BottomSheet/ReviewSheet/ReviewBottomSheetTitle.tsx
@@ -1,9 +1,0 @@
-function ReviewBottomSheetTitle() {
-  return (
-    <h1 className="w-full h-auto flex items-center text-xl bg-white font-semibold leading-6">
-      리뷰작성
-    </h1>
-  );
-}
-
-export default ReviewBottomSheetTitle;
